+++ src/main/java/com/raoulvdberge/refinedstorage/tile/grid/TileGrid.java
@@ -3,7 +3,6 @@ package com.raoulvdberge.refinedstorage.tile.grid;
 import com.raoulvdberge.refinedstorage.api.network.grid.GridType;
 import com.raoulvdberge.refinedstorage.api.network.grid.IGrid;
 import com.raoulvdberge.refinedstorage.apiimpl.network.node.NetworkNodeGrid;
-import com.raoulvdberge.refinedstorage.container.ContainerGrid;
 import com.raoulvdberge.refinedstorage.gui.GuiBase;
 import com.raoulvdberge.refinedstorage.gui.grid.GuiGrid;
 import com.raoulvdberge.refinedstorage.tile.TileNode;
@@ -72,16 +71,6 @@ public class TileGrid extends TileNode<NetworkNodeGrid> {
         t.getNode().setProcessingPattern(v);
         t.getNode().clearMatrix();
         t.getNode().markDirty();
-
-        t.getWorld().getMinecraftServer()
-            .getPlayerList()
-            .getPlayers()
-            .stream()
-            .filter(player -> player.openContainer instanceof ContainerGrid && ((ContainerGrid) player.openContainer).getTile() != null && ((ContainerGrid) player.openContainer).getTile().getPos().equals(t.getPos()))
-            .forEach(player -> {
-                ((ContainerGrid) player.openContainer).initSlots();
-                ((ContainerGrid) player.openContainer).sendAllSlots();
-            });
     }, (initial, p) -> GuiBase.executeLater(GuiGrid.class, GuiBase::initGui));
     public static final TileDataParameter<Integer, TileGrid> PROCESSING_TYPE = IType.createParameter((initial, p) -> GuiBase.executeLater(GuiGrid.class, GuiBase::initGui));